Fixed race condition
Race condition will happen when there is a subscription but there is a close() in emit callback, and mostly occur in redis cluster. Most tests have close() in emit function callback.
The emit callback is wrapped in setImmediate, and it will be called after publish I/O in the event loop. That means the close() in the emit callback will be called before the next I/O which trigger subscription event. Not a good solution, but work and we seldom to call close() to shutdown the mqemitter-redis.